Digishop/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
PyJWT==1.7.1
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.2
